Don't embed pointer-like structs
Some handle and pointer types are in the metadata as wrapper structs
with one member (e.g. HBITMAP, LPWSTR). The previous change to support
embedded structs would turn all of these into properties with the name
"Value". This would cause it to generate invalued AutoHotkey code, and
is generally unhelpful.

Now, these are treated as pointer-types.

This does create an additional problem for structs like
NMCUSTOMSPLITRECTINFO[1] that have more than one copy of a struct
embedded.

[1] https://learn.microsoft.com/en-us/windows/win32/api/commctrl/ns-commctrl-nmcustomsplitrectinfo