Use an ISignatureTypeProvider for field decoding
Previously I was doing this manually; I had assumed that because AHK is
relatively simple in its type system compared to C, that would be
sufficient. Turns out the .NET people designed System.Reflection the
way they did on purpose, and this earlier decision was going to make
parsing method parameters basically impossible.