Only use StructLayout packing size if LayoutKind = Explicit
Previously, if the struct layout's declared packing size was nonzero,
we always used it. However, this resulted in incorrectly calculated
packing sizes and thus incorrectly generated offsets in cases where the
layout size was nonzero but the struct layout was not explicit.
example, CHOOSECOLORW and CHOOSECOLORA have a nonzero packing size but
a layout kind of sequential:

[StructLayout(LayoutKind.Sequential, Pack = 1)]

This caused us to generate the struct with an alignment of 1 byte, and
the struct was missing several padding blocks as a result.

Now, we only use the declared packing size if it's nonzero AND the
layout kind is LayoutKind.Explicit