Fix TypeName for pointers from pseudo-primitives
Previously, if we encountered a pseudo-primitive, the name of the
pointed-to value would be incorrect when the type was parsed (it would
be the name of the pseudo-primitive itself, not the name of the wrapped
primitive). For example, HWND's value is a void*, so the pointed-to
name should be "void", but it was HWND.

Now, this is fixed.

Additionally, Pointer types now store information about the pointed-to
type in the field previously reserved for array types. That's been
renamed to UnderlyingType to reflect its new responsibilities.