Detect Buffers and emit as primitive Pointers
Some method parameters are in practice Buffers whose size is determined
by another parameter in the argument list. See for example pbRandomData
in Cryprography.SystemPrng [1][2] (and ProcessPrng).

Here, the [MemorySize(BytesParamIndex = 1)] attribute is telling us to
look at argument 1 (cbRandomData) for the size of the buffer pointed to
by pbRandomData. Note also theat pbRandomData shows up as a pointer to
a byte, which is not what we want to send to DllCall - we need a 64-bit
(pointer-sized) integer. See also [3].

So now we detect these attributes and emit the arguments as primitive
pointers. It's probably possible to use the information here to emit a
line of AHK code that checks the length of the buffer, but I'm not
doing that here.

[1] public unsafe static extern BOOL SystemPrng([Out][MemorySize(BytesParamIndex = 1)] byte* pbRandomData, [In] UIntPtr cbRandomData);
[2] https://learn.microsoft.com/en-us/windows/win32/SecCNG/systemprng
[3] https://github.com/microsoft/win32metadata/issues/284