Use star syntax for DllCall pointer->primitive arguments
Previously, these were emitted as raw pointers (using the "ptr" DllCall
type). This required the user to first aquire a pointer to the variable
and then call the wrapper function. In practice this meant wrapping
primitives in Buffers all over the place, which is the kind of pattern
that this project exists to eliminate.

Now, when emitting functions, for arguments which are pointers to
primitives other than Void, we'll use DllCall's star-syntax [1] to pass
them to Windows. This also allows callers to pass VarRefs in cases
where the function they are calling will modify the variable.

[1] https://www.autohotkey.com/docs/v2/lib/DllCall.htm#asterisk