Fix String argument detection
A previous commit [1] changed the way that we store data for pointers,
which broke string detection for methods specifically. Previously, for
pseudo-primitives like PWSTR and PSTR, we incorrectly stored the name
of the pseudo-primitive type in the field intended for the name of the
type of the pointed-to value, and checked that do detect strings.

Now we check the TypeDefinition directly.

[1] c843fc20a847aaa367e1857d42319cfb279a5b98