Fix return value detection
The previous fix [1] to pointed-to type detection means that now, void*
pointers look like void* pointers. When checking to see if a method has
a return value, we previously checked to see if the return value's type
was void, but if the return type is a Pointer, this is legit.

Now, we check that the return type is both a primitive AND a void.

[1] c843fc20a847aaa367e1857d42319cfb279a5b98