Respect PreserveSig and related attributes
The generator now checks and respects the following method attributes:
    - [DllImport(..., PreserveSig = true/false)]
    - [CanReturnMultipleSuccessValues]
    - [CanReturnErrorsAsSuccess]

In practice, this means that we now check the return values of methods
that return HRESULTs and throw if they aren't E_SUCCESS (0), unless
either [CanReturnMultipleSuccessValues] or [CanReturnErrorsAsSuccess]
is present. Those indicate that the method *can* return error values as
success, though in practice these are mostly on COM classes which are
not currently supported. Note also that we have no way to know which
error codes might be returned as success values, so we just punt error
checking to callers. This seems to be what Rust and C# do as well.

This also required a change to field definition parsing because HRESULT
is a pseudo-primitive. There's now a specific fieldKind for it, and we
document it using the name HRESULT for clarity.

    /**
     * Error code returned by the operation.
     * @type {HRESULT}
     */
    ApiStatus {
        get => NumGet(this, 24, "int")
        set => NumPut("int", value, this, 24)
    }

See also:
[1] https://github.com/microsoft/win32metadata/issues/1315#issuecomment-1279835954
[2] https://learn.microsoft.com/en-us/dotnet/api/system.runtime.interopservices.dllimportattribute.preservesig?view=net-7.0#remarks