Use EntryPoint instead of MethodDefinition Name in DllCalls
Previously, I'd assumed that the name of the MethodDefinition was also
the name of the method in the corresponding Dll (that is, the value you
would send to GetProcAddress). This assumption was never safe, though
coincidentally it does hold for almost every method in the metadata.

Now, we use MethodDefinition.Import.Name as the entry point into the
Dll, as the ECMA335 specification specifies.

See https://github.com/holy-tao/AhkWin32Projection/issues/29#issue-3479546381