Only emit x64 and arm64 struct variants
While investigating another issue [1], I discovered that the generator
was incorrectly emitting x86 structs, sometimes overwriting x64 and
arm64 structs. This could lead to problems, as 32-bit AutoHotkey is not
supported. Now we only check the [SupportedArchitecture] attribute and
only emit structs where it has either x64 or arm64. This is still
imperfect, as some structs have distinct x64 and arm64 variants [2].
But it's good enough for the 99% case, and AHK doesn't have a great way
to conditionally include structs based on architecture.

[1] https://github.com/holy-tao/AhkWin32Projection/issues/32
[2] https://github.com/holy-tao/AhkWin32Projection/blob/main/Windows/Win32/System/Kernel/SLIST_HEADER.ahk