Resolve nested TypeReferences with their parents
When resolving TypeReferences, we just searched the entire module for a
matching fully qualified name. This works 99% of the time, but in some
cases there are multiple types with identical names in the a namespace
because they're for different architectures, e.g. SLIST_HEADER.

Now we allow callers to create FieldSignatureProvider objects with a
TypeDefinition context in which to search when resolving TypeReferences

See https://github.com/holy-tao/AhkWin32Projection/issues/35