Check for reserved names in Ahk Types
There's a struct in Kernal called STRING that conflicts with the native
type String. This would cause a duplicate declaration load-time error
if it were ever #Included in a script. Keep a short list of reserved
words and check struct names against them, prepend "Win32" to the name
if we find a conflict.

See https://github.com/holy-tao/AhkWin32Projection/pull/42