Wrap handles returned from functions
Apis will now import handle types they need and wrap returned handles.
This commit also includes logic for parsing the [DoNotRelease] and
[IgnoreIfReturn] attributes - if [DoNotRelease] is present, the handle
that is returned is marked unowned, if [IgnoreIfReturn] is present, we
generate a guard statement that checks the result of the DllCall to see
if we should return an invalid handle.