Fix: Don't filter for field count
Missed this one earlier. Filtering for field count was not a reliable
way to find types we don't want to project; much better to check their
types. Now we do that.

This also caused a bug in Api generation where Apis which did not have
constants would be incorrectly excluded. This included namespaces like
InteractionContext.