The sameful return of IsPseudoPrimitive (kind of)
It's more selective now. We will flatten types if they meet all of the
following requirements:
1. They have the [NativeTypedef] attribute
2. They have exactly one member
3. They are not handles (at least, we can't tell from the metadata)

This means that types like PSTR and PWSTR are emitted as pointers, but
handles like HWND and HINSTANCE are emitted as handle types. We should
be a lot less likely to catch stray structs in this net. Documentation
also uses the Windows type of the field, even if it's a primitive under
the hood.