Introduce extensions
The generator will search the "extensions" subdirectory of the metadata
directory for .yml and .yaml files containing extensions and read them
into a static "singleton" in Program called Extensions. Extension files
are parsed into a dictionary keyed by the fully-qualified names of the
types to which the extensions apply; the value is a list of the objects
holding the actual code.

We don't actually use these yet