Use [Documentation] attr for link if not set in ApiDetails
Some types don't have entries in Win32Docs, and others don't have help
links. But the metadata includes a [Documentation] attribute with links
to MSDN pages for most types and methods. This is especially pronounced
wrt COM interfaces and methods - almost none of them have documentation
in Win32Docs but almost all of them have the [Documentation] attribute.

If a type or method doesn't have documentation or lacks a link but does
have the [Documentation] attribute, use it.