Resolve references to types in external assemblies
Previously, when we encountered a TypeReference with a ResolutionScope
of Assembly, we simply assumed that it was a pointer and created a fake
FieldInfo of type (Primitive, "ptr"). While it is true in practice that
to use all of these types we ultimately must use pointers, generation
has become increasingly complex and we need more information about the
pointed-to type (for example, to know whether a pointer to an external
type should be resolved as ptr* or ptr**). The most notable of these is
System.Guid, but there may be more I'm not aware of.

When we encounter a TypeReference with a ResolutionScope of Assembly,
load the assembly and search it to get the actual TypeDefinition. Throw
errors if we cannot resolve a reference instead of defaulting.