Marshal pointers to primitives based on the input parameter type
Pointers to primitives in method signatures (int*, float*) were always
passed to DllCall with an asterisk type, to allow users to pass VarRefs
and get responses back. But, as the issue notes, it is impossible to
pass raw pointers or NULL for optional arguments.

For pointers to primitives, type check the input parameter. If passed a
VarRef, pass it to DllCall as a * type, otherwise, use a raw ptr

Fixes https://github.com/holy-tao/AhkWin32Structs-Generator/issues/18