Scan metadata directory for .winmd files
Previously, we just looked for "windows.win32.winmd" in the directory.
But WinRT has many .winmd files and they'll need to go somewhere. This
changes the main generation loop to scan the metadata directory for
.winmd files.

It also changes the shape of the version.ini file to use the assembly
version instead of metadata version