Handle nonstandard [RAIIFree] for handle return values
Handle return values can have an explicit [RAIIFree] attribute, either
because they're special and need to be closed a specific way or because
they're generic type like HANDLE and don't have a default destructor.
We now handle this case by calling DefineProp[1] on the handle before
returning it and binding "Free" to the destructor. See also [2].

[1] https://www.autohotkey.com/docs/v2/lib/Object.htm#DefineProp
[2] https://github.com/holy-tao/AhkWin32Projection/issues/58#issuecomment-3584613523