refactor: consolidate error checking codegen
Since we might need to free resources before throwing an error when a
function call fails, we need that logic to all be in one place. Also,
we can't use the HRESULT return type[1] since that tells the DllCall
(or ComCall) itself to throw before we get the chance to free anything.
Now we don't.

[1] https://www.autohotkey.com/docs/v2/lib/DllCall.htm#HRESULT