Use TypeDefinition reader to parse parameter information
I changed the type resolution logic a while ago to bundle a FieldDef's
MetadataReader with it's type information, but most places that need to
access type information didn't use this MetadataReader. This was fine
when the only assembly being read was Win32, but with WDK and in the
future when WinRT assemblies are added, we need to use the correct
MetadataReader for the assembly that the type is into avoid exceptions.