Don't wrap com output parameters
In most cases we try to wrap function return values in the appropriate
type to make things easier on the user. This causes problems for com
pointers though, because the declared type of the parameter will almost
always be too general - IUnknown, IInspectable, etc. While this is
technically correct, it makes things harder in most case because the
caller, who knows which interface they queried for / which class they
just activated, needs to convert from the generic interface to the more
specific one.

When output parameters have the [ComOutPtr] attribute, just return them
as raw pointers. Callers will need to wrap values themselves - but this
is easier with a primitive than with an interface proxy.