Handle WinRT interfaces and cross-assembly inheritance
WinRT interfaces use the implemented interfaces property the way it was
designed - instead of listing the interface whose vtable the current
one extends, it lists the interfaces that you can query the current one
for safely. All WinRT interfaces implicitly extend IUnknown.

Changes to handle that, and to handle the cross-assembly inheritance of
WinRT interfaces to Win32's IUnknown.