Fix assembly loading for WinRT interfaces
Previously assemblies were loaded in two places - the Program.cs loop,
and in FieldSignatureDecoder. FieldSignatureDecoder has a cache for the
assemblies it loads, but Program.cs would unload assemblies when it was
done processing them, causing some pointers in the cache to become
invalid. This wasn't a problem with just Win32 for obvious reasons, and
I guess we got lucky with Wdk because it's loaded before Win32, so
FieldSignatureDecoder loads it ad hoc before Program does. However with
WinRT this becomes a problem.

Don't unload assemblies in Program, fix the way they're referenced at
codegen time.