Prevent name conflicts with type names
WinRT has enough parameters with names that conflict with type names
that it's not possible to keep a hardcoded list. Now, load the names of
all types in an assembly into a trie and use that to see if we're going
to have a conflict. This is a little overly broad, but will prevent
issues.