Account for method parameter differences in WinRT vs Win32
WinRT methods all use the HRESULT / Output pointer convention for their
parameter orders, but obnoxiously, this isn't actually in the metadata.
Instead, parameter 0 (the output parameter) is the logical return value
and is actually the *last* parameter to the method. So, we need to tack
it on to the end of the parameter list artificially, and we also need
to create a "synthetic" paramter zero of type Primitive / HRESULT.

Ultimately, the format of the method's paramter list follows the Win32
convention. Parameter 0 is the return value, and the rest of the method
params are it's actual parameters in the order we send them to DllCall.