Fix: Pin temporary String handles
With WinRT methods we regularly encounter heap corruption crashes when
creating emporary handles and allowing __Delete to run on its own time.
I haven't seen any evidence that this is also a problem with BSTRs, but
we might as well pin those as well while we're at it.

When creating temporary handles for caller-supplied strings in wrapper
methods, pin a reference to them, guaranteeing that they never fall out
of scope before the method ends and therefore aren't destructed until
after the DllCall / ComCall is finished.