Fix Class -> Interface method overload deduplication logic
WinRT Class methods and COM interface methods used slightly different
logic to deduplicate overloaded methods, and COM interface methods did
not consider the [Overload] attribute (which is only present in WinRT
metadata). Related, if a class had multiple methods with the same name,
but they were declared+ on different interfaces, the class would not
deduplicate the overload names, and even if it did, the assumption that
the name of the interface method is the same as the name of the class
method that wraps it would break.

This makes a series of related changes to unify the overload name
deduplication logic, make COM interfaces aware of [Overload], and force
the class-level wrappers to use the method name as declared on the COM
interface instead of assuming that it matches the name on the class.