Support generic interfaces
WinRT class methods that return generic interfaces will now correctly
use the parameterized interface id when querying the base interface for
the generic.

Generics are handled on the AHK side by prepending the list of generic
arguments to the interface's constructor. Interfaces that take generic
args receive them as references to functions that, when called with the
pointer to wrap, return objects of the generic arg's type. Typically
this just means Interface.Call, but in the case of HSTRINGS might be an
anonymous function that creates an HSTRING from the value of the return
type. It helps that nearly every generic type is boxed in an
IPropertyValue.

Interface methods that return generics return them by calling the bound
generic factory function with one argument, the pointer to wrap.