Precompute parameterized interface IDs
Parameterized interfaces [1] use special Guids called parameterized
interface ids (PIIDs) which are computed [2] based on an algorithm from
the WinRT spec. This introduces a new tool to precompute the PIIDs that
appear in the WinRT metadata. This allows the generator to treat them
like compile-time (code-generation, whatever) constants.

It is possible to compute these at runtime; Microsoft has an API [3]
for language projections to do this, but it requires parsing metadata,
which somewhat defeats the purpose of a language projection (Compiled
languages don't have this problem).

[1] https://learn.microsoft.com/en-us/uwp/winrt-cref/winrt-type-system#parameterized-types
[2] https://learn.microsoft.com/en-us/uwp/winrt-cref/winrt-type-system#guid-generation-for-parameterized-types
[3] https://learn.microsoft.com/en-us/windows/win32/api/roparameterizediid/nf-roparameterizediid-rogetparameterizedtypeinstanceiid